requirements.txt
antlr4-python3-runtime==4.9
apispec==4.0.0
bokeh==2.2.3
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorcet==2.0.2
dask==2.30.0
datashader==0.11.1
datashape==0.5.2
distributed==2.30.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-PyMongo==2.3.0
flask-smorest==0.25.0
fsspec==0.8.4
gunicorn==20.0.4
HeapDict==1.0.1
holoviews==1.14.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
llvmlite==0.35.0
locket==0.2.0
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.9.1
msgpack==1.0.0
multipledispatch==0.6.0
numba==0.52.0
numpy==1.19.4
packaging==20.7
pandas==1.1.5
panel==0.10.2
param==1.10.0
partd==1.1.0
Pillow==8.0.1
psutil==5.7.3
pyarrow==2.0.0
pyct==0.4.6
pymongo==3.11.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
pyviz-comms==0.7.6
PyYAML==5.3.1
requests==2.25.0
scipy==1.5.4
shortuuid==1.0.1
six==1.15.0
sortedcontainers==2.3.0
tblib==1.7.0
toolz==0.11.1
tornado==6.1
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.26.2
webargs==6.1.1
Werkzeug==1.0.1
xarray==0.16.2
zict==2.0.0
